requirements.txt
coreapi==2.3.3
dj-database-url==0.5.0
Django==2.0.6
djangorestframework==3.8.2
gunicorn==19.8.1
Markdown==2.6.11
psycopg2==2.7.5
Pygments==2.2.0
python-dateutil==2.7.3
Unipath==1.1
whitenoise==3.3.1
